requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
fastapi
uvicorn
pillow
numpy
scikit-learn
python-multipart
dvc
pytest
